[EDELIVERY-8424] [EDELIVERY-8239] Fix issue with pMode not showing in the UI [PR #2520]

+++ Domibus-MSH/src/main/java/eu/domibus/core/pmode/multitenancy/MultiDomainPModeProvider.java
@@ -274,7 +274,7 @@ public class MultiDomainPModeProvider extends PModeProvider {
     }
 
     @Override
-    public byte[] getPModeFile(int id) {
+    public byte[] getPModeFile(long id) {
         return getCurrentPModeProvider().getPModeFile(id);
     }
